Always skip the S3 HEAD check from the local client.
For cold builds, the latency hit of adding round-trips is worse than
the cost of just uploading the extra files, and for warm builds the
in-memory cache does the job better. We might consider extending the
default daemon lifetime, or persisting the "known to exist on remote"
set to disk to make it more effective.