Improve llama's robustness to $HOME being unset. (#32)
In a Bazel build environment, $HOME is not set by default and it can be
hard to get Basel to set it in all circumstances. To avoid the need for
a wrapper script, break the dependency on $HOME by using the go-homedir
package, which will file the home directory correctly even when $HOME
isn't set.

The AWS session setup also depends on $HOME, but we can work around
that by presetting the environment variables to the default shared
configuration paths.