Forward preprocessor definitions to the remote compiler.
When doing partial local preprocessing, a remote clang compilation
still needs the preprocessor definitions from the command line so
that the remaining preprocessing can be performed correctly. Partial
preprocessing is necessary to prevent the remote compiler issuing
spurious warnings from macro expansion that would be suppressed in
a local compilation.

This means when spawning a compilation, we should always use either
`LocalArgs` or `RemoteArgs`, and the need to separately preserve
unknown arguments and conditionally restore preprocessor definitions
goes away.

This updates #37.