Add warning filtering.
Some build systems build multiple projects and have different ways to
detect the right compilation flags, so it can be difficult to ensure
that the local build sends the right warning flags to the remote compilers.

Add a LLAMACC_FILTER_WARNINGS configuration option that causes llamacc to
filter out the specified warning flags (in both the enabled and disabled
variants). this can be a useful hack when experimenting with llamacc in
a complex build system.

This updates #40.